requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
gunicorn==21.2.0
idna==3.6
load-dotenv==0.1.0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
